requirements.txt
bottle==0.12.18
PyYAML==5.2